+++ karaboFAI/config.py
@@ -174,6 +174,14 @@ class _Config(dict):
             "MODULE_SHAPE": [256, 256],
             "PIXEL_SIZE": 0.5e-3,
         },
+        "JungFrauPR": {
+            "REDIS_PORT": 6383,
+            "PULSE_RESOLVED": True,
+            "REQUIRE_GEOMETRY": False,
+            "NUMBER_OF_MODULES": 2,
+            "MODULE_SHAPE": [512, 1024],
+            "PIXEL_SIZE": 0.075e-3,
+        },
         "JungFrau": {
             "REDIS_PORT": 6380,
             "PULSE_RESOLVED": False,
@@ -301,6 +309,27 @@ class _Config(dict):
             "CENTER_X": 570,
             "PHOTON_ENERGY": 9.3,
         },
+        "JungFrauPR": {
+            "SERVER_ADDR": "10.253.0.53",
+            "SERVER_PORT": 4501,
+            # Source name from bridge not yet known.
+            "SOURCE_NAME_BRIDGE": [
+                "FXE_XAD_JF1M/CAL/APPEND",
+                "FXE_XAD_JF1M/DET/RECEIVER-1:daqOutput",
+                "FXE_XAD_JF1M/DET/RECEIVER-2:daqOutput",
+                "FXE_XAD_JF500K/DET/RECEIVER:display",
+            ],
+            # Source name from files not yet known.
+            "SOURCE_NAME_FILE": [
+                "FXE_XAD_JF1M/DET/RECEIVER-1:daqOutput",
+                "FXE_XAD_JF1M/DET/RECEIVER-2:daqOutput",
+            ],
+            "AZIMUTHAL_INTEG_RANGE": [0.05, 0.4],
+            "SAMPLE_DISTANCE": 2.0,
+            "CENTER_Y": 512,
+            "CENTER_X": 1400,
+            "PHOTON_ENERGY": 9.3,
+        },
         "JungFrau": {
             "SERVER_ADDR": "10.253.0.53",
             "SERVER_PORT": 4501,
@@ -346,7 +375,9 @@ class _Config(dict):
             "SERVER_ADDR": '10.253.0.140',
             "SERVER_PORT": 4511,
             "SOURCE_NAME_BRIDGE": [
-                'SCS_CDIDET_DSSC/CAL/APPEND_CORRECTED'],
+                'SCS_CDIDET_DSSC/CAL/APPEND_CORRECTED',
+                'SCS_CDIDET_DSSC/CAL/APPEND_RAW',
+            ],
             "SOURCE_NAME_FILE": ['SCS_CDIDET_DSSC/CAL/APPEND_CORRECTED'],
             "GEOMETRY_FILE": osp.join(osp.dirname(osp.abspath(__file__)),
                                       'geometries/dssc_geo_june19.h5'),